Add get by id and get all methods Add corresponding API gateway path and methods Move DB logic to a separate module
