+++ deploy.sh
@@ -1,0 +1,3 @@
+. ./.venv/bin/activate
+./node_modules/aws-cdk/bin/cdk deploy
+deactivate
